게임 프로토타입 완성 아마......
지난 커밋에서 안되던 데미지 적용 문제 해결방법 생각나서 해결하고, 각종 UI구현, 플레이어가 적과 닿으면 플레이어 체력 감소 구현, Game Start, Game Over Panel 구현해서 플레이어 체력이 0이하가 되면 GameOver 구현,
ReStart버튼도 구현하여 시작 신으로 loadScene했음. 맵도 약간 꾸밈.
아무튼 프로토타입임. 언젠간 다시 할게요...

+++ Assets/myGame/Script/EnemyData.cs
@@ -12,7 +12,7 @@ public class EnemyData : ScriptableObject
     public int Hp { get { return hp; } }
 
     [SerializeField] private int damage;
-    public int Damage { get { return damage; } }
+    public float Damage { get { return damage; } }
 
     [SerializeField] private float moveSpeed;
     public float MoveSpeed { get { return moveSpeed; } }
